pyproject.toml
[tool.poetry]
name = "fitagent-nutrition-coach"
version = "1.0.0"
description = "AI-powered nutrition coaching agent for Fetch.ai Agentverse"
authors = ["FitAgent Team"]

[tool.poetry.dependencies]
python = "^3.9"
uagents = "^0.12.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
flask = "^2.3.0"
flask-cors = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
